refactor :home screen
