+++ composeApp/src/commonMain/kotlin/com/cairosquad/evolvefit/ui/screen/favorites/content/componants/DeleteFavouriteSnackBar.kt
@@ -1,0 +1,2 @@
+package com.cairosquad.evolvefit.ui.screen.favorites.content.componants
+
